refactored the person methods
